pyproject.toml
[tool.poetry]
name = "karaokay-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
numpy = "^1.26.4"
python-pptx = "^0.6.23"
python-dotenv = "^1.0.1"
openai = "^1.12.0"
nltk = "^3.8.1"
stopwordsiso = "^0.6.1"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
